+++ quill/include/quill/Logger.h
@@ -154,7 +154,8 @@ public:
     std::byte* write_buffer =
       thread_context->spsc_queue<QUILL_QUEUE_TYPE>().prepare_write(static_cast<uint32_t>(total_size));
 
-    if constexpr (QUILL_QUEUE_TYPE == detail::QueueType::BoundedNonBlocking)
+    if constexpr ((QUILL_QUEUE_TYPE == detail::QueueType::BoundedNonBlocking) ||
+                  (QUILL_QUEUE_TYPE == detail::QueueType::UnboundedDropping))
     {
       if (QUILL_UNLIKELY(write_buffer == nullptr))
       {
@@ -163,7 +164,12 @@ public:
         return;
       }
     }
-    else if constexpr (QUILL_QUEUE_TYPE == detail::QueueType::BoundedBlocking)
+    else if constexpr (QUILL_QUEUE_TYPE == detail::QueueType::UnboundedNoMaxLimit)
+    {
+      assert(write_buffer && "UnboundedNonBlocking will always allocate and get new space");
+    }
+    else if constexpr ((QUILL_QUEUE_TYPE == detail::QueueType::BoundedBlocking) ||
+                       (QUILL_QUEUE_TYPE == detail::QueueType::UnboundedBlocking))
     {
       while (write_buffer == nullptr)
       {