+++ quill/include/quill/Quill.h
@@ -29,8 +29,8 @@ namespace quill
 {
 
 /** Version Info **/
-constexpr uint32_t VersionMajor{2};
-constexpr uint32_t VersionMinor{10};
+constexpr uint32_t VersionMajor{3};
+constexpr uint32_t VersionMinor{0};
 constexpr uint32_t VersionPatch{0};
 constexpr uint32_t Version{VersionMajor * 10000 + VersionMinor * 100 + VersionPatch};