+++ quill/include/quill/detail/ThreadContext.h
@@ -37,7 +37,8 @@ public:
   explicit ThreadContext(QueueType queue_type, uint32_t default_queue_capacity, uint32_t initial_transit_event_buffer_capacity)
     : _transit_event_buffer(initial_transit_event_buffer_capacity)
   {
-    if (queue_type == QueueType::Unbounded)
+    if ((queue_type == QueueType::UnboundedBlocking) ||
+        (queue_type == QueueType::UnboundedNoMaxLimit) || (queue_type == QueueType::UnboundedDropping))
     {
       _spsc_queue.emplace<UnboundedQueue>(default_queue_capacity);
     }
@@ -81,9 +82,13 @@ public:
    * @return A reference to the generic single-producer-single-consumer queue
    */
   template <QueueType queue_type>
-  QUILL_NODISCARD_ALWAYS_INLINE_HOT std::conditional_t<queue_type == QueueType::Unbounded, UnboundedQueue, BoundedQueue>& spsc_queue() noexcept
+  QUILL_NODISCARD_ALWAYS_INLINE_HOT std::conditional_t<(queue_type == QueueType::UnboundedBlocking) || (queue_type == QueueType::UnboundedNoMaxLimit) ||
+                                                         (queue_type == QueueType::UnboundedDropping),
+                                                       UnboundedQueue, BoundedQueue>&
+  spsc_queue() noexcept
   {
-    if constexpr (queue_type == QueueType::Unbounded)
+    if constexpr ((queue_type == QueueType::UnboundedBlocking) ||
+                  (queue_type == QueueType::UnboundedNoMaxLimit) || (queue_type == QueueType::UnboundedDropping))
     {
       return std::get<UnboundedQueue>(_spsc_queue);
     }
@@ -97,9 +102,13 @@ public:
    * @return A reference to the generic single-producer-single-consumer queue const overload
    */
   template <QueueType queue_type>
-  QUILL_NODISCARD_ALWAYS_INLINE_HOT std::conditional_t<queue_type == QueueType::Unbounded, UnboundedQueue, BoundedQueue> const& spsc_queue() const noexcept
+  QUILL_NODISCARD_ALWAYS_INLINE_HOT std::conditional_t<(queue_type == QueueType::UnboundedBlocking) || (queue_type == QueueType::UnboundedNoMaxLimit) ||
+                                                         (queue_type == QueueType::UnboundedDropping),
+                                                       UnboundedQueue, BoundedQueue> const&
+  spsc_queue() const noexcept
   {
-    if constexpr (queue_type == QueueType::Unbounded)
+    if constexpr ((queue_type == QueueType::UnboundedBlocking) ||
+                  (queue_type == QueueType::UnboundedNoMaxLimit) || (queue_type == QueueType::UnboundedDropping))
     {
       return std::get<UnboundedQueue>(_spsc_queue);
     }