+++ quill/include/quill/detail/spsc_queue/UnboundedQueue.h
@@ -117,6 +117,15 @@ public:
           " max_bounded_queue_capacity: " + std::to_string(max_bounded_queue_capacity)});
       }
 
+      if constexpr ((QUILL_QUEUE_TYPE == detail::QueueType::UnboundedBlocking) ||
+                    (QUILL_QUEUE_TYPE == detail::QueueType::UnboundedDropping))
+      {
+        // we reached the unbounded queue limit of 2147483648 bytes (~2GB) we won't be allocating
+        // anymore and instead return nullptr to block
+        return nullptr;
+      }
+
+      // else the UnboundedNonBlocking queue has no limits and will keep allocating additional 2GB queues
       capacity = max_bounded_queue_capacity;
     }